refactor(pipeline): fix imports for IDE debugging
- Change relative imports to simple imports for direct execution
- Rename logging.py to logger.py to avoid conflict with built-in module
- Add unescape_string function to handle Pretalx newline formatting
- All modules can now be run directly in IDE debugger
